Replace singledispatch for abstract spaces by custom class
Advantage is that the new class overloads operators for direct sums and
scaling similar to the concrete spaces, and that it has built-in
`partial` functionality.

It's less general than singledispatch, but should be enough for our
case.

Check's should still only require `callable(x)` instead of
`ininstance(x, AbstractSpace)`, for more flexibility.