Avoid modifying or aliasing operator arguments
Returning arrays that alias the argument (e.g. for the Identity)
operator can lead to subtle and hard to debug bugs when using in-place
modification. E.g. TikhonovCG failed on L2Generic domains since self.dir
and self.g_res end up being identical, but are both modified in-place.

Making copies may be overly pessimistic in many cases, but the API is
much harder to screw up if the caller can assume operators to not modify
arguments, and to always return an array that does not share data with
anything alive, i.e. the caller "owns" both argument and return value.