Functionals: allow multiple ways of implementing gradients
For coherence with the Operator interface, using a .linearize method is
most natural. Also, functionals involving operators will need to call
op.linearize when computing the gradient, anyway, so most likely it will
be cheap to return the functional value as well.

On the other hand, minimizing a functional often does not need the
absolute value, so having to call .linearize each time can become
cumbersome.

This commit implements a .gradient method that computes only the
gradient, deferring the actual computation to the ._gradient abstract
method. If that is not implemented, it falls back to ._linearize.

Since there are now to options to compute the value (_eval and
_linearize), gradient (_gradient and _linearize) and linearization
(_linearize and _eval+_gradient), the public methods try both. Users are
free to implement only the necessary ones.

Also, the _gradient method is implemented for all currently existing
functionals.

That's way more commit message than code, by the way.